Add hook support for optional user scripts
If you put scripts in a hook directory:

* `/etc/puppetfactory/hooks/create/`
* `/etc/puppetfactory/hooks/delete/`

Then each script will be executed in turn, passing the username as an
argument, any time a user is created or destroyed. Output is silently
logged to the puppetfactory logfile (`/var/log/puppetfactory` by
default).

The hook directory location is configurable via `puppetfactory.yaml`.