Improve logging and resilience
This logs errors instead of falling over. The log will contain all user
creation/deletion/etc events. It will rescue errors, log them, and
return a useful error message instead of just crashing.

On creation, it will stop at the first error, but on deletion it will
keep going and try to remove each artifact that it can. This should
leave less crap lying around on the system after an error.

Some logic was simplified, but it should not have drastically changed
behaviour.