Add simple local/remote CLI
This just adds a handful of quick commands to list, add, remove users.
It uses the REST API, which means that this can be run from the
puppetfactory machine itself (defaults to sending commands to localhost)
or from any machine which can reach it.

The REST API is terrifyingly unsecured and that needs to be fixed soon.