Slightly more sophisticated user auth
This changes how user authentication is handled. Now students don't need
to log in to view the users page at all. Instead, when they create a
user, they're prompted for a session ID that will be provided by the
instructor. Once they create a user, that username is saved in a session
cookie.

This username is not currently used, but when we decide on the UI/UX
complications of displaying only the current user on the users tab, this
could easily be used for this purpose.

The instructor can elevate privileges by clicking the button in the top
right and logging in. The button becomes visible when mousing over.

Because this is somewhat precipitated by the need for an instructor
dashboard, the beginnings of that are conflated with this change.

The next project will be to build the dashboard and actually flesh out
that tab. (Currently, the tab can be disabled by setting it to false in
the config file.)