Add a personalized user view
This will put a blurb at the top of the users page with the current
user's information, including a link to their console node group.

This also refactors the user listing to not gather costly data, so the
page is overall much much faster.

Finally, there's a link to select any user to display, because we're not
asking users to actually log in to Puppetfactory yet.