Adds a link to the users's current test
This just puts a link in the student area to the HTML copy of their
current Rspec test output. It's not foolproof, because it depends on the
instructor to remember to update the current running test. If they leave
it on 'summary' then the output is overwhelming.

Ideas for a better UX?