Use an actual shell instead of a .bashrc
This makes login more reliable and doesn't leave an extra bash lying
around for each login. It also means that login FAILS instead of just
leaving the user logged in to the master in case something goes wrong.

This also has the benefit that you no longer need a real home directory
if you want to investigate other options for mapping the puppetcode
directory. It works just fine with the user's home set to /tmp.

TRAINTECH-687 #resolved #time 15m