Refactored Puppetfactory to a cleaner architecture
This mostly maintains status quo, but refactors quite a bit. The base
Puppetfactory is now very limited. It will only create shell accounts
and allow users to log into them.

To add more features, you enable plugins. For example, we'd likely use a
list of these plugins for Puppetizing:

```
:plugins:
- :Certificates
- :Classification
- :Docker
- :Logs
- :Dashboard
- :CodeManager
- :ShellUser
```

The only *new* features this currenly adds are the `puppetfactory
configprint` command which displays the currently loaded configuration,
and the `Logs` plugin, which adds a view of the Puppetfactory logfile in
a tab for admin users.

The configuration is rather different though. Rather than repeating the
defaults in many places, I just assign sane defaults, then the module
just overrides any that it needs. This restructures the configuration a
bit so there are fewer Boolean options (since that's handled by enabling
a plugin now) and a few other more generic ways of describing options.

The new configuration and the plugin system is documented in the README,
and the `pltraining/puppetfactory` module should manage it properly.

Note: this depends on Abalone v0.3.2, for more straightforward
embedding.

TRAINTECH-709 #resolved #time 5d