Increase Thin timeout
The Thin middleware server was timing out (30s by default) when the
classroom got loaded down. This caused the user creation to sometimes be
incomplete, and would lead to a poor user experience when they got raw
HTML squished into a popup dialog box!

The better solution is a progress bar, but that's out of scope for the
current sprint.

TRAINTECH-1392 #resolved #time 6h